python/mydbpkg/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mydbpkg"
version = "0.1.0"
description = "Exercise package for MySQL CRUD functions"
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["mysql-connector-python"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
